database-component/requirements.txt
paho-mqtt==1.5.1
mongoengine==0.21.0
pymongo==3.11.0